# app-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ChapApp-Api
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /users/checkexisteduser:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Check existence of user
      # used as the method name of the controller
      operationId: checkExistedUser
      parameters:
        - name: username
          in: query
          description: username
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResponseString"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
...
  /users/getdata:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get data of user
      # used as the method name of the controller
      operationId: getdata
      parameters:
        - name: token
          in: query
          description: Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UserData"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
  /users/getall:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get all user
      # used as the method name of the controller
      operationId: getalluser
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResponseString"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
  /users/getuser:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get the user's friend list
      # used as the method name of the controller
      operationId: getusercon
      parameters:
        - name: token
          in: query
          description: Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResponseString"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
  /users/setstatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: Set status for user
      # used as the method name of the controller
      operationId: setstatus
      parameters:
        - name: userStatus
          in: body
          description: Status
          required: true
          schema:
            $ref: "#/definitions/UserStatus"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/Response"
...
  /user/getstatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get status of user
      # used as the method name of the controller
      operationId: getstatus
      parameters:
        - name: token
          in: query
          description: Token
          required: true
          type: string
        - name: username
          in: query
          description: username
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResponseString"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
...
definitions:
...
  Response:
    properties:
      status:
        type: integer
      message:
        type: string
...
  ChatItem:
    properties:
      id:
        type: number
      username:
        type: string
      content:
        type: string
      time:
        type: number  
  CongroupInfo:
    properties:
      conId:
        type: number
      username:
        type: string
      groupId:
        type: number
      groupname:
        type: string 
      unread:
        type: integer
      createUser:
        type: boolean
      lastedmsg:
        type: object
        $ref: "#/definitions/ChatItem"
  UserData:
    properties:
      status:
        type: integer
      user:
        type: string
      userStatus:
        type: string
      listConGroup:
        type: array
        items:
          $ref: "#/definitions/CongroupInfo"
  ResponseString:
    properties:
      status:
        type: integer
      value:
        type: string
  UserStatus:
    properties:
      status:
        type: string
      token:
        type: string 
...